backend/requirements.txt
tensorflow>=2.12.0
pillow==9.2.0
flask==2.2.3
flask-cors==3.0.10
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
opencv-python==4.7.0.72
numpy ==1.23.5